Absolute horror it's been.
Added a function that simplifies updating popup.js with newer bookmarks but event listener is being ignored.
So the bookmarks are being registered but the event listener looking for changes in chrome.storage isn't being triggered for some reason. This means that the popup.html doesn't get updated as the bookmarks are added to chrome storage. Will try to fix soon.